+++ src/strategies/long_call.rs
@@ -5,6 +5,7 @@
 #![allow(clippy::indexing_slicing)]
 
 use super::base::{BreakEvenable, Positionable, StrategyType};
+use crate::analytics::probability::VolatilityAdjustment;
 use crate::backtesting::results::{SimulationResult, SimulationStatsResult};
 use crate::chains::OptionChain;
 use crate::error::{
@@ -28,7 +29,6 @@ use crate::strategies::base::lower_break_even;
 use crate::strategies::base::price_gap;
 use crate::strategies::delta_neutral::DeltaNeutrality;
 use crate::strategies::probabilities::core::ProbabilityAnalysis;
-use crate::strategies::probabilities::utils::VolatilityAdjustment;
 use crate::strategies::utils::OptimizationCriteria;
 use crate::strategies::{
     BasicAble, DeltaAdjustment, Strategable, Strategies, StrategyConstructor, Validable,